moved examples out
