+++ Changelog.md
@@ -1,0 +1,5 @@
+# Changelog
+
+## 0.2.0
+
+- packaged with uv
